simplify soy project upon review
