+++ boilerplates/webpack-with-soy/package.json
@@ -1,6 +1,5 @@
 {
 	"name": "metal-modal",
-	"publicName": "Modal",
 	"version": "0.0.0",
 	"description": "My awesome Metal.js project",
 	"license": "BSD",
@@ -9,13 +8,13 @@
 		"node": ">=0.12.0",
 		"npm": ">=3.0.0"
 	},
-	"main": "src/index.js",
+	"main": "src/Modal.js",
 	"files": [
 		"src",
 		"test"
 	],
 	"scripts": {
-		"build": "npm run soy && node scripts/build.js",
+		"build": "npm run soy && webpack",
 		"soy": "metalsoy --soyDeps node_modules/metal*/src/**/*.soy node_modules/clay*/src/**/*.sory"
 	},
 	"keywords": [
@@ -31,17 +30,7 @@
 		"babel-loader": "^7.1.1",
 		"babel-preset-es2015": "^6.24.1",
 		"babel-preset-metal": "^4.1.0",
-		"css-loader": "^0.28.4",
-		"detect-port": "^1.2.1",
-		"fs-extra": "^4.0.1",
-		"inquirer": "^3.2.2",
 		"metal-tools-soy": "^4.2.6",
-		"node-sass": "^4.5.3",
-		"react-dev-utils": "^3.1.1",
-		"sass-loader": "^6.0.6",
-		"style-loader": "^0.18.2",
-		"webpack": "^3.5.4",
-		"webpack-dev-server": "^2.7.1",
-		"webpack-visualizer-plugin": "^0.1.11"
+		"webpack": "^3.5.4"
 	}
 }
