+++ boilerplates/webpack-with-soy/src/Modal.js
@@ -1,5 +1,3 @@
-import './modal.scss';
-
 import templates from './Modal.soy.js';
 import Component from 'metal-component';
 import Soy from 'metal-soy';
