--- boilerplates/webpack-with-soy/src/modal.scss
@@ -1,6 +1,0 @@
-/**
- * Custom style definitions
- */
-.Modal {
-	
-}
