+++ boilerplates/webpack-with-soy/webpack.config.js
@@ -1,0 +1,30 @@
+const webpack = require('webpack');
+const path = require('path');
+const fs = require('fs');
+const url = require('url');
+
+// Make sure any symlinks in the project folder are resolved
+const appDirectory = fs.realpathSync(process.cwd());
+function resolveApp(relativePath) {
+  return path.resolve(appDirectory, relativePath);
+}
+
+module.exports = {
+  entry: './src/Modal.js',
+  output: {
+    path: resolveApp('./build'),
+    filename: 'modal.js',
+    library: 'Modal'
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader'
+        }
+      }     
+    ]
+  }
+};
